refactor searchbar component, remove book actions debugger
